requirements.txt
PySide6==6.7.2
Pillow>=10.0.0
matplotlib>=3.8.0
fastapi>=0.99,<0.100; python_version < "3.10"
fastapi>=0.110.0; python_version >= "3.10"
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic<2; python_version < "3.10"
starlette<0.28; python_version < "3.10"

# Optional OCR backends (install at least one)
pix2tex>=0.1.2
pix2text>=1.1.2
